Revert '500 on socket_resume'.
Apparently socket_resume returns when the call is bridged, not when we disconnect the socket.
This causes the call to terminate as soon as it is connected.

(This was already committed as 7e15f1a51c1d73b878cf88f550ffb194f991dbec, so re-reverting change in 99f6d9c48db9c58fe1a5ec857667b696f7ed3d89 )